chore: move service-response to model file
