--- src/data/service-responses/index.ts
@@ -1,5 +1,0 @@
-import { LessonServiceResponse } from "./lesson.service-response";
-
-export type {
-  LessonServiceResponse
-};
